// src/app/core/models/shop.model.ts
export class Shop {
    id: number;
    user_id: number;
    agent_user_id: number;
    country_id: number;
    state_id: number;
    city_id: number;
    name: string;
    logo: string;
    cover_photo: string;
    shopkeeper_nrc: string;
    nrc: string;
    lat: string;
    long: string;
    address: string;
}